added mean absolute percentage error
